sigo avanzando con IO_GEN_SLEEP (a revisar entre todos)

+++ cpu/src/instrucciones.h
@@ -6,4 +6,4 @@ char* obtener_nombre_instruccion(t_instruccion* instruccion);
 
 // FUNCIONALIDAD DE LAS INSTRUCCIONES
 void ejecutar_set(char* registro, void* valor_recibido);
-void ejecutar_dormirIO(int nombreInterfaz, int unidadesDeTiempo);
+void ejecutar_dormirIO(int nombreInterfaz, uint8_t unidadesDeTiempo);
